Added wl-clip-pesists to the clipboard manager list
Normally, when we copy something on Wayland and then close the application we copied from, the copied data (e.g. text) disappears and we cannot paste it anymore. It is very frustrating when I copy something from a clipboard and the copied data disappears after I close the application. It effects peoples work flow. So we can use something called as wl-clip-persists (https://github.com/Linus789/wl-clip-persist) which will preserve the data in the clipboard even after the application from where we copied the data is closed.