Update Binds.md
By default binds trigger on press of a key (as opposed to on release). There are some exceptions to this, notably tofi with the exec dispatcher.  This is problematic since tofi gets on the screen in a single frame, so the bottleneck for the user will be this anomaly in trigger behavior, which the l flag fixes. I might not have worded it really well, but i think its important to document this issue