Configuring/Uncommon-tips-&-tricks: Improved User-Experience of "Alt tab behaviour" (#1306)
* Configuring/Uncommon-tips-&-tricks: Improved User-Experience of "Alt tab behaviour"

The previous implementation executed the `focuswindow` and `alterzorder`
dispatchers before the footclient process terminated. This created a
timing issue that is particularly noticable during window focusing
across workspaces.

For example: A user on workspace 2 uses the alt-tab command to focus one
out of several windows on workspace 1 which was not previously focused
when the user was working on workspace 1. The focus is correctly updated
but Hyprland fails to draw the active border around the newly focused
window.

This is fixed by updating the focus after the footclient has terminated.

* Alt Tab Behaviour uses XDG_RUNTIME_DIR instead of /tmp to support multiple users on the same machine